% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_from_stage.R
\name{age_from_stage}
\alias{age_from_stage}
\alias{mpm_to_lx}
\alias{mpm_to_px}
\alias{mpm_to_hx}
\alias{mpm_to_mx}
\title{Calculate age-specific traits from a from matrix population model}
\usage{
mpm_to_lx(matU, startLife, N)

mpm_to_px(matU, startLife, N)

mpm_to_hx(matU, startLife, N)

mpm_to_mx(matU, matR, startLife, N)
}
\arguments{
\item{matU}{The survival component of a matrix population model (i.e. a
square projection matrix reflecting survival-related transitions; e.g.
progression, stasis, and retrogression)}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life.}

\item{N}{Maximum age to which age-specific traits will be calculated.}

\item{matR}{The reproductive component of a matrix population model (i.e. a
square projection matrix reflecting transitions due to reproduction; either
sexual, clonal, or both)}
}
\value{
A vector of length \code{N+1}
}
\description{
These functions use age-from-stage decomposition methods to calculate
age-specific survivorship (lx), survival probability (px), mortality hazard
(hx), or reproduction (mx) from a matrix population model. A detailed
description of these methods can be found in sections 5.3.1 and 5.3.2 of
Caswell (2001).
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

# age-specific survivorship
mpm_to_lx(matU, startLife = 1, N = 20)

# age-specific survival probability
mpm_to_px(matU, startLife = 1, N = 20)

# age-specific mortality hazard
mpm_to_hx(matU, startLife = 1, N = 20)

# age-specific fecundity
mpm_to_mx(matU, matF, startLife = 1, N = 20)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\seealso{
\code{\link{lifetable_convert}}
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>

Patrick Barks <patrick.barks@gmail.com>
}
