% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dentropy.R
\name{dentropy}
\alias{dentropy}
\title{Calculates Demetrius' entropy}
\usage{
dentropy(matU, matF, matC = FALSE, startLife = 1, nSteps = 1000)
}
\arguments{
\item{matU}{A matrix containing only survival-dependent processes (e.g. progression,
stasis, retrogression).}

\item{matF}{A matrix containing only sexual reproduction, with zeros
elsewhere.}

\item{matC}{A matrix containing only clonal reproduction, with zeros
elsewhere. If not provided, it defaults to a matrix with all zeros.}

\item{startLife}{The first stage at which the author considers the beginning
of life in the life cycle of the species. It defaults to the first stage.}

\item{nSteps}{A cutoff for the decomposition of age-specific survival ('lx'), and when pertinent,
for age-specific sexual reproduction ('mx') and age-specific clonal reproduction ('cx'). This allows
excluding mortality and fertility plateaus. See function 'qsdConverge' for more information. When not
specified, this argument defaults to 100.}
}
\value{
Returns an estimate of Demetrius' entropy. When both 'matF' and 'matC' are provided,
it outputs Demetrius' entropy for sexual reproduction only ('Fec'), for clonal reproduction only ('Clo'),
and for both types of reproduction together ('FecClo').
}
\description{
This function calculates Demetrius' entropy from a matrix population model
}
\note{
%% ~~further notes~~
}
\examples{
matU <- matrix (c(0, 0, 0, 0, 0.5, 0, 0, 0, 0, 0.3, 0, 0, 0, 0, 0.1, 0.1), nrow = 4, byrow = TRUE)
matF <- matrix (c(0, 0, 5, 10, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0), nrow = 4, byrow = TRUE)
matC <- matrix (c(0, 0, 0, 0, 0, 0, 0, 1, 0, 2, 0, 0, 0, 0, 0, 0), nrow = 4, byrow = TRUE)

dentropy(matU, matF, matC, nSteps=5)
dentropy(matU, matF, matC, nSteps=10)
dentropy(matU, matF, matC, nSteps=100)

#' @import MASS
}
\references{
L. Demetrius. 1978. Adaptive value, entropy and survivorship
curves. Nature 275, 213 - 214. doi:10.1038/275213a0
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
