% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_k.R
\name{entropy_k}
\alias{entropy_k}
\title{Calculate Keyfitz's entropy from a trajectory of age-specific survivorship}
\usage{
entropy_k(lx, trapeze = FALSE)
}
\arguments{
\item{lx}{Survivorship trajectory (a vector of monotonically-declining values
in the interval [0,1])}

\item{trapeze}{A logical argument indicating whether the composite trapezoid
approximation should be used for approximating the definite integral.}
}
\value{
Keyfitz's life table entropy.
}
\description{
This function calculates Keyfitz's entropy from a vector of age-specific
survivorship (lx).
}
\examples{
lx <- 0.8^(0:20)

entropy_k(lx)
entropy_k(lx, trapeze = TRUE)

}
\references{
Keyfitz, N. (1977) Applied Mathematical Demography. New York:
  Wiley.
}
\author{
Owen R. Jones <jones@biology.sdu.dk>

Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>
}
