% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie_collapse.R
\name{is_leslie}
\alias{is_leslie}
\title{Check whether a matrix is a Leslie matrix}
\usage{
is_leslie(A)
}
\arguments{
\item{A}{a matrix to be checked}
}
\value{
a logical value indicating whether the matrix is a Leslie matrix
}
\description{
Checks whether a matrix is a Leslie matrix by verifying that the matrix
satisfies three conditions: (1) the matrix has non-negative elements, (2) the
elements on the subdiagonal are between 0 and 1, and (3) the sum of the birth
rates is positive.
}
\examples{
A <- matrix(c(0.1, 1.2, 1.1, 
              0.0, 0.2, 0.0, 
              0.0, 0.0, 0.3), nrow = 3, byrow = TRUE)
is_leslie(A) # true
A <- matrix(c(0.1, 1.2, 1.1, 
              0.1, 0.2, 0.1, 
              0.2, 0.3, 0.3), nrow = 3, byrow = TRUE)
is_leslie(A) # false?
}
\author{
Richard A. Hinrichsen <rich@hinrichsenenvironmental.com>
}
