% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kEntropy.R
\name{kEntropy}
\alias{kEntropy}
\title{Calculates Keyfitz' entropy}
\usage{
kEntropy(matU, startLife = 1, nSteps = 1000, trapeze = FALSE,
  trimlx = 0.05)
}
\arguments{
\item{matU}{A matrix containing only survival-dependent processes (e.g. progression,
stasis, retrogression).}

\item{startLife}{The first stage at which the author considers the beginning
of life in the life cycle of the species. It defaults to the first stage.}

\item{nSteps}{A cutoff for the decomposition of age-specific survival ('lx'), and when pertinent,
for age-specific sexual reproduction ('mx') and age-specific clonal reproduction ('cx'). This allows
excluding mortality and fertility plateaus. See function 'qsdConverge' for more information. When not
specified, this argument defaults to 100.}

\item{trapeze}{A logical argument indicating whether the trapezoidal
approximation should be used for approximating the definite integral.}

\item{trimlx}{A numerical argument determining how lx should be trimmed. 
For example, a value of 0.05 means that lx values are retained only if they 
are greater than or equal to 0.05. NB: 0 values in lx mean that Keyfitz 
entropy cannot be estimated.}
}
\value{
Returns an estimate of Keyfitz' life table entropy based on an lx
(survivorship) vector obtained from matU
}
\description{
This function calculates Keyfitz' entropy from an lx
(survivorship) vector with even intervals derived from a matrix population model.
}
\examples{

matU <- matrix (c(0, 0, 0, 0, 0.6, 0, 0, 0, 0, 0.4, 0, 0, 0, 0, 0.7, 0.1), nrow = 4, byrow = TRUE)
kEntropy(matU, nSteps=100)
kEntropy(matU,trapeze=FALSE)

matU <- matrix (c(0.2, 0, 0, 0, 0.3, 0.4, 0.1, 0, 0.1, 0.1, 0.2, 0.3, 0, 0.2, 0.6, 0.5), nrow = 4, byrow = TRUE)
kEntropy(matU, nSteps = 10)
kEntropy(matU, nSteps = 20)
kEntropy(matU, nSteps = 100)
kEntropy(matU, nSteps = 100, trapeze=TRUE)

}
\references{
%% ~~references~~
}
\author{
Owen R. Jones <jones@biology.sdu.dk>

Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>
}
