% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leslie_collapse.R
\name{Leslie_collapse}
\alias{Leslie_collapse}
\title{Aggregate a Leslie matrix}
\usage{
Leslie_collapse(A, m)
}
\arguments{
\item{A}{a Leslie matrix}

\item{m}{the dimensionality of the desired aggregated matrix}
}
\value{
a list including the following elements:
\item{B}{the aggregated matrix}
\item{W}{the weight matrix}
\item{Ak}{the original (or expanded) Leslie matrix raised to the k power}
\item{S}{the partitioning matrix}
\item{n}{the size of the original (or expanded) Leslie matrix}
\item{m}{the size of the aggregated matrix}
\item{k}{the quotient of the original size divided by the aggregated size}
\item{EFF}{the effectiveness of aggregation}
}
\description{
Takes a Leslie matrix and aggregates it to a desired
dimensionality m using least squares weights equal to the stable age
distribution. The output includes the aggregated matrix, the weight matrix,
the original (or expanded) Leslie matrix raised to the k power, the
partitioning matrix, the size of the original (or expanded) Leslie matrix,
the size of the aggregated matrix, the quotient of the original size divided
by the aggregated size, and the effectiveness of aggregation.
}
\examples{
data(leslie_mpm1)
A <- leslie_mpm1$matU + leslie_mpm1$matF
Leslie_collapse(A, 4)

}
\author{
Richard A. Hinrichsen <rich@hinrichsenenvironmental.com>
}
