% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leslie_mpm1}
\alias{leslie_mpm1}
\title{Example Leslie matrix population model (MPM)}
\format{
A list with two elements:
\describe{
  \item{matU}{The survival-related component of the MPM.}
  \item{matF}{The sexual reproduction component of the MPM.}
}
}
\usage{
leslie_mpm1
}
\description{
An example Leslie matrix population model (MPM) used for demonstration and
testing purposes.
}
\keyword{datasets}
