% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeExpectancy.R
\name{lifeExpectancy}
\alias{lifeExpectancy}
\title{Calculate life expectancy from a matrix population model}
\usage{
lifeExpectancy(matU, startLife = 1)
}
\arguments{
\item{matU}{A matrix containing only survival-dependent processes (growth,
stasis, shrinkage).}

\item{startLife}{Index of the first stage at which the author considers the
beginning of life. Defaults to 1.}
}
\value{
Returns life expectancy. If \code{matU} is singular (often indicating
  infinite life expectancy), returns \code{NA}.
}
\description{
Applies Markov chain approaches to obtain mean life expectancy from a matrix
population model.
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

lifeExpectancy(matU)
lifeExpectancy(matU, startLife = 2)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\author{
Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <hcaswell@whoi.edu>
}
