% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTimeRepEvents.R
\name{lifeTimeRepEvents}
\alias{lifeTimeRepEvents}
\title{Calculate the timing of lifetime reproductive events}
\usage{
lifeTimeRepEvents(matU, matR, startLife = 1)
}
\arguments{
\item{matU}{A square matrix containing only survival-related transitions
(i.e. progression, stasis, retrogression).}

\item{matR}{A square matrix containing only reproduction-related transitions
(either sexual, clonal, or both; i.e. \code{matF}, \code{matC}, or
\code{mat + matC}).}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life. Defaults to 1.}
}
\value{
A list with four elements:
\item{p}{probability of achieving reproductive maturity}
\item{La}{mean age at maturity (in the same units as the matrix population
model sampling periodicity)}
\item{meanLifeExpectancy}{mean life expectancy conditional on entering the
life cycle at the value of \code{startLife}}
\item{remainingMatureLifeExpectancy}{Life expectancy from mean maturity,
calculated as \code{meanLifeExpectancy - La}. This value can be negative
because both mean life expectancy and mean age at maturity are means of their
respective distributions, and their distributions can indeed overlap}
}
\description{
Applies Markov chain approaches to decompose various moments along the
age-specific trajectory of reproduction of individuals in a matrix population
model, including the probability of achieving maturity, age at first
reproduction, mean life expectancy conditional on maturity, and life
expectancy for mature individuals.
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

lifeTimeRepEvents(matU, matF, startLife = 1)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\author{
Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <hcaswell@whoi.edu>

Owen R. Jones <jones@biology.sdu.dk>
}
