% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longevity.R
\name{longevity}
\alias{longevity}
\title{Calculate longevity from a matrix population model}
\usage{
longevity(matU, start = 1, xmax = 1000, lx_crit = 0.01)
}
\arguments{
\item{matU}{The survival component of a matrix population model (i.e. a
square projection matrix reflecting survival-related transitions; e.g.
progression, stasis, and retrogression)}

\item{start}{Index of the first stage at which the author considers the
beginning of life. Defaults to 1.}

\item{xmax}{The maximum age to which survivorship will be calculated.
Defaults to 1000.}

\item{lx_crit}{The critical proportion to calculate longevity with respect to
(a value between 0 and 1). Defaults to 0.01.}
}
\value{
Returns longevity, the integer age at which expected survivorship
  falls below \code{lx_crit}. If survivorship doesn't reach \code{lx_crit} by
  \code{xmax}, returns \code{NA} and prints a warning message.
}
\description{
Calculate longevity (the age a which survivorship falls to some critical
proportion) from a matrix population model
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

longevity(matU)
longevity(matU, lx_crit = 0.05)
longevity(matU, start = 3, xmax = 50, lx_crit = 0.05)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968

  Morris, W. F., and D. F. Doak. (2003) Quantitative Conservation Biology:
  Theory and Practice of Population Viability Analysis. Sinauer Associates,
  Sunderland, Massachusetts, USA
}
\author{
Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <hcaswell@whoi.edu>
}
