% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLifeTable.R
\name{makeLifeTable}
\alias{makeLifeTable}
\title{Generate a life table from a matrix population model}
\usage{
makeLifeTable(matU, matF = NULL, matC = NULL, startLife = 1,
  nSteps = 1000)
}
\arguments{
\item{matU}{A square matrix containing only survival-related transitions
(i.e. progression, stasis, retrogression).}

\item{matF}{(Optional) A square matrix containing only sexual
reproduction-related transitions.}

\item{matC}{(Optional) A square matrix containing only clonal
reproduction-related transitions.}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life. Defaults to 1.}

\item{nSteps}{Number of time steps for which the life table will be
calculated. Time steps are in the same units as the matrix population model
(see MatrixPeriodicity column in metadata of COM(P)ADRE). Defaults to 1000.}
}
\value{
A \code{list} containing 7-19 named elements. List elements are
  numeric vectors of length \code{nSteps}, unless otherwise noted.\cr\cr
  Always includes the following:
\tabular{lll}{
  \code{x} \tab \tab age\cr
  \code{lx} \tab \tab survivorship to start of age class x\cr
  \code{dx} \tab \tab proportion of original cohort dying in interval [x, x+1)\cr
  \code{qx} \tab \tab force of mortality at age x\cr
  \code{Lx} \tab \tab survivorship to middle of age class x\cr
  \code{Tx} \tab \tab proportion of original cohort alive at age x and beyond\cr
  \code{ex} \tab \tab remaining life expectancy at age x\cr
}
If \code{matF} is provided, also includes:
\tabular{lll}{
  \code{mx} \tab \tab per-capita rate of sexual reproduction at age x\cr
  \code{lxmx} \tab \tab expected number of sexual offspring per original
  cohort member produced at age x\cr
  \code{R0Fec} \tab \tab <scalar> net reproductive rate (sexual)\cr
  \code{TcFec} \tab \tab <scalar> mean generation time (sexual)\cr
}
If \code{matC} is provided, also includes:
\tabular{lll}{
  \code{cx} \tab \tab per-capita rate of clonal reproduction at age x\cr
  \code{lxcx} \tab \tab expected number of clonal offspring per original
  cohort member produced at age x\cr
  \code{R0Clo} \tab \tab <scalar> net reproductive rate (clonal)\cr
  \code{TcClo} \tab \tab <scalar> mean generation time (clonal)\cr
}
If both \code{matF} and \code{matC} are provided, also includes:
\tabular{lll}{
  \code{mxcx} \tab \tab per-capita rate of total reproduction (sexual + clonal) at age x\cr
  \code{lxmxcx} \tab \tab expected number of total offspring (sexual + clonal) per original
  cohort member produced at age x\cr
  \code{R0FecClo} \tab \tab <scalar> net reproductive rate (total reproduction)\cr
  \code{TcFecClo} \tab \tab <scalar> mean generation time (total reproduction)\cr
}
}
\description{
This function uses age-from-stage decomposition methods to generate a life
table from a matrix population model. A detailed description of these methods
can be found in section 5.3 of Caswell (2001).
}
\examples{
matU <- rbind(c(0.0, 0.0, 0.0, 0.0),
              c(0.5, 0.0, 0.0, 0.0),
              c(0.0, 0.3, 0.0, 0.0),
              c(0.0, 0.0, 0.1, 0.1))

matF <- rbind(c(0.0, 0.0, 5.0, 8.1),
              c(0.0, 0.0, 0.0, 0.0),
              c(0.0, 0.0, 0.0, 0.0),
              c(0.0, 0.0, 0.0, 0.0))

matC <- rbind(c(0.0, 0.0, 0.1, 0.2),
              c(0.0, 0.0, 0.2, 1.3),
              c(0.0, 0.0, 0.0, 0.0),
              c(0.0, 0.0, 0.0, 0.0))

makeLifeTable(matU, startLife = 1, nSteps = 100)
makeLifeTable(matU, matF, startLife = 1, nSteps = 100)
makeLifeTable(matU, matF, matC, startLife = 1, nSteps = 100)
}
\references{
Caswell, H. (2001) Matrix Population Models: Construction, Analysis, and
Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Caswell, H. (2006) Applications of Markov chains in demography. pp. 319-334
in A.N. Langville and W.J. Stewart (editors) MAM2006: Markov Anniversary
Meeting. Boson Books, Raleigh, North Caroline, USA

Jones, O.R. et al. (2014) Diversity of ageing across the tree of life.
Nature, 505(7482), 169–173
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>
}
