% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixElementPerturbation.R
\name{matrixElementPerturbation}
\alias{matrixElementPerturbation}
\title{A function to perform element perturbation of a matrix population model.}
\usage{
matrixElementPerturbation(matU, matF, matC = NULL, pert = 0.001)
}
\arguments{
\item{matU}{The U matrix (processes related to survival, growth and retrogression).}

\item{matF}{The F matrix (sexual reproduction processes).}

\item{matC}{The C matrix (clonal reproduction processes).}

\item{pert}{Perturbation parameter.}
}
\value{
%% ~Describe the value returned
}
\description{
A function to perform element perturbation of a matrix population model.
}
\note{
%% ~~further notes~~
}
\examples{


}
\references{
%% ~~references~~
}
\author{
Roberto Salguero-Gomez <r.salguero@sheffield.ac.uk>
}
