% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpm1}
\alias{mpm1}
\title{Example matrix population model (MPM)}
\format{
A list with two elements:
\describe{
  \item{matU}{The survival-related component of the MPM}
  \item{matF}{The sexual reproduction component of the MPM}
}
}
\usage{
mpm1
}
\description{
An example matrix population model (MPM) used for demonstration and testing.
Consists of five stage classes: 'seed', 'small', 'medium', 'large', and
'dormant'.
}
\keyword{datasets}
