% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_rep.R
\name{pace_rep}
\alias{pace_rep}
\title{Calculate shape of reproduction over age}
\usage{
pace_rep(rep, xmin = NULL, xmax = NULL, fertTable = FALSE)
}
\arguments{
\item{rep}{Either 1) a numeric vector describing reproduction over age (mx); 2) a 
\code{data.frame} / \code{list} with one column / element titled 'mx' 
describing a reproduction over age, optionally a column / element 'x' containing 
age classes (each element a number representing the age at the start of the
class); 3) a list containing two elements: 'matU', a U matrix (the survival 
component of a matrix population model, i.e. a square projection matrix reflecting 
survival-related transitions, e.g. progression, stasis, and retrogression) and
'matF', and F matrix(the reproduction component of a matrix projection model, 
i.e. a square projection matrix of the same dimension as matU reflecting 
those transitions describing sexual reproduction); 4) a \code{CompadreMat} 
object (\code{RCompadre-package})containing a matrix population model in 
the format described in the \code{CompadreMat} class.
In the case of 1 and 2 where x is not supplied, the function will assume
age classes starting at 0 with steps of 1 year.
In the case of 3 and 4, an age-based reproduction schedule will be generated from 
a stage-based matrix using the \code{makeLifeTable} function of 
\code{RCompadre-package}. 
In all cases where x ends at maximum longevity, \code{mx[which.max(x)]} 
should equal 0, however it is possible to supply partial reproduction schedules.}

\item{xmin, xmax}{The minimum and maximum age repectively over which to evaluate
shape. If not given, these default to \code{min(x)} and \code{max(x)} 
respectively.}

\item{fertTable}{logical determining whether to return the fertility table 
used to calculate pace, including standardised measures of age, fertility
and standardised fertility.}
}
\value{
the pace of reproduction, equal to the age of the mother at which an
  average baby is born. This is the birth equivalent to life expectancy, 
  calculated using the cumulative reproduction.
}
\description{
Calculates a 'shape' value of distribution of reproduction over age by comparing
the area under a cumulative reproduction curve (over age) with the area 
under a cumulative function describing constant reproduction.
}
\examples{
mx <- c(0, 0, 0.3, 0.4, 0.5, 0.6)
pace_rep(mx)

}
\author{
Iain Stott <iainmstott@gmail.com>
}
