% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_surv.R
\name{pace_surv}
\alias{pace_surv}
\title{Calculate shape of survival over age}
\usage{
pace_surv(surv, survTable = FALSE)
}
\arguments{
\item{surv}{Either 1) a numeric vector describing a survival curve (lx), or
  2) a \code{data.frame} / \code{list} with one column / element titled 'lx'
  describing a survival curve, optionally a column / element 'x' containing
  age classes (each element a number representing the age at the start of the
  class). 

  If x is not supplied, the function will assume age classes starting at 0
  with time steps of 1 unit. If x begins at 0 then \code{lx[1]} should equal
  1. If x ends at maximum longevity, then \code{lx[which.max(x)]} should
  equal 0; however it is possible to supply partial survivorship curves.}

\item{survTable}{Logical to determine whether a survival table should be
returned (see details).}
}
\value{
the pace of survival, equal to life expectancy at birth. If 
  \code{survTable = TRUE}, a list containing the pace of survival, and 
  a survival table which contains the parameters required to calculate pace:
  Lx (midpoint survivorshop), Tx (total number of age categories left for
  individuals surviving to beginning of age interval), ex (life expectancy
  at the start of the age category). Note that life expectancy at other ages 
  (especially at age of maturity) may also be considered as measures of 
  the pace of survival.
}
\description{
Calculates a 'shape' value of survival lifespan inequality by comparing the 
area under a survival curve (over age) with the area under a constant 
survival function.
}
\examples{
# exponential decline in lx yields shape = 0
lx <- 0.7^(0:20)
pace_surv(lx)

}
\author{
Iain Stott <iainmstott@gmail.com>
}
