% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLifeCycle.R
\name{plotLifeCycle}
\alias{plotLifeCycle}
\title{Plot the life cycle diagram based on the matrix model}
\usage{
plotLifeCycle(A, title = "my life cycle", shape = "egg",
  fontsize = 10, nodefontsize = 12, edgecol = "grey")
}
\arguments{
\item{A}{The A matrix of a matrix population model}

\item{title}{A title for the plot}

\item{shape}{The shape to be used for the stages of the diagram. Any node
shape accepted by graphViz is acceptable.}

\item{fontsize}{Size of the font used in the diagram.}

\item{nodefontsize}{Size of the font used in the node part of the diagram.}

\item{edgecol}{Colour of the arrows in the diagram.}
}
\value{
Produces a plot consisting of a life cycle diagram, based on the A
matrix.
}
\description{
Plots the life cycle diagram illustrated by a matrix model. It processes the
matrix model and passes the information to the graphViz functionality of
DiagrammeR. See http://rich-iannone.github.io/DiagrammeR/.
}
\examples{

\dontrun{
M1 <- matrix(c(0.00, 0.69, 0.00, 0.10, 0.00, 0.88, 0.35, 0.00, 0.79),nrow=3)
colnames(M1) <- 1:3
plotLifeCycle(M1)
} 


}
\author{
Owen R. Jones <jones@biology.sdu.dk>
}
