% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_vectors.R
\name{pop_vectors}
\alias{pop_vectors}
\title{Derive a hypothetical set of population vectors corresponding to a
time-series of matrix population models}
\usage{
pop_vectors(A, start = "stable.stage")
}
\arguments{
\item{A}{A list of MPMs (i.e. square population projection matrices).}

\item{start}{Method to derive the first population vector in the series.
Either `stable.stage` to use the stable stage distribution of the mean
matrix as the starting vector, `uniform` to use a uniform starting vector
(all elements equal), or `random` to use a randomly-generated starting
vector. Defaults to `stable.stage`.}
}
\value{
A list of population vectors
}
\description{
Derive a hypothetical set of population vectors (i.e. population size
distributions across stages) given a time-series of matrix population models
(MPMs), by taking the stable stage distribution of the mean matrix as the starting
vector (or optionally, a uniform or random starting vector), and deriving
subsequent vectors through recursive population projection.
}
\examples{
# generate list of matrices
matA_l <- replicate(5, matrix(runif(9), 3, 3), simplify = FALSE)

# calculate corresponding population vectors
pop_vectors(matA_l)
pop_vectors(matA_l, start = "uniform")
pop_vectors(matA_l, start = "random")

}
\seealso{
Other {transformation}: 
\code{\link{mpm_collapse}()},
\code{\link{mpm_split}()},
\code{\link{mpm_standardize}()},
\code{\link{nameStages}()},
\code{\link{plot_life_cycle}()},
\code{\link{repro_stages}()},
\code{\link{standard_stages}()}
}
\author{
Patrick Barks <patrick.barks@gmail.com>
}
\concept{{transformation}}
