% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsdConverge.R
\name{qsdConverge}
\alias{qsdConverge}
\title{Cut-off of quasi-stable distribution}
\usage{
qsdConverge(matU, conv = 0.05, startLife = 1, nSteps = 1000)
}
\arguments{
\item{matU}{The survival-dependent matrix (a subset of the A matrix), See
Caswell 2001, or other references below.}

\item{conv}{Departure of the projection population vector from convergence
to the stationary st/age distribution. E.g. this value should be 0.05 if the
user wants to obtain the time step when the stage distribution is within 5
percent of the stationary st/able distribution.}

\item{startLife}{The stage in the life cycle of the organism, as represented
by the matrix population model, where the user considers 'the begining of
life'. This is useful to discard pre-established stages such as permanent
seedbanks..}

\item{nSteps}{Number of time steps for which the projection will be made. In
most cases the default of 1000 is sufficient to reach stationary st/age
distribution. This is on the same units as the matrix population model - see
MatrixPeriodicity in metadata of COMPADRE/COMADRE.}
}
\value{
Returns a numeric value indicating the time step at which
quasi-stationary stage distribution has been reached.
}
\description{
Calculates the time in the projection of a cohort through a matrix population model at which
a defined quasi-stationary stage distribution is reached in \code{\link{Rage}}
}
\details{
Size or stage-based matrix population models (i.e. Lefkovitch models) and some age-based
matrix population models (i.e. Leslie models with a last class ≥ final age value) are
typically parameterised with a stasis loop in the
largest/most-developed stage (e.g. adult survival). The assumption
of constancy in vital rates for individuals in that last stage typically results in flat
mortality and fertility plateaus. These plateaus may result in mathematical artefacts
when examining age-specific patterns dervied from age-from-stage matrix
decompositions (Caswell 2001). The Quasi-stationary Stage Distribution (QSD)
can be used to circunvent this problem. The QSD is the stage distribution
that is reached some time before the ultimate stable stage distribution (SSD, the
normalised right eigenvector of the transition matrix). With this approach, the user
can ask what is the time step in the projectin at which the cohort
approximates its stable stage distribution with a given convergence tolerance level (e.g. 95%).
This metric allows the user to use only age-based information from before this point. See the
online supplementary information of Jones et al. (2014) for further details.
}
\examples{
matU <- matrix(c(0.000, 0.691, 0.000, 0.000, 0.000, 0.880, 0.000, 0.000, 0.790),nrow=3)
qsdConverge(matU, 0.05)
}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
978-0878930968

Jones, O.R. et al. (2014) Diversity of ageing across the tree of life.
Nature, 505(7482), 169–173

Salguero-Gómez R. Implications of clonality for ageing research.
Evolutionary Ecology DOI 10.1007/s10682-017-9923-2
}
\author{
Hal Caswell <h.caswell@uva.nl>

Owen Jones <jones@biology.sdu.dk>

Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
\keyword{methods}
