% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R0.R
\name{R0}
\alias{R0}
\title{Calculate net reproductive value}
\usage{
R0(matU, matR, startLife = 1)
}
\arguments{
\item{matU}{A matrix containing only survival-related transitions (i.e.
progression, stasis, retrogression).}

\item{matR}{A matrix containing only reproduction-related transitions (either
sexual, clonal, or both; i.e. \code{matF}, \code{matC}, or \code{mat +
matC}).}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life. Defaults to 1.}
}
\value{
Returns the net reproductive value.
}
\description{
Calculate net reproductive value from a matrix population model.
}
\examples{
matU <- rbind(c(0.0, 0.0, 0.0, 0.0),
              c(0.5, 0.0, 0.0, 0.0),
              c(0.0, 0.3, 0.0, 0.0),
              c(0.0, 0.0, 0.1, 0.1))

matF <- rbind(c(0, 0, 5, 9),
              c(0, 0, 0, 0),
              c(0, 0, 0, 0),
              c(0, 0, 0, 0))

R0(matU, matF)
R0(matU, matF, startLife = 2)
}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>
}
