% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rage-package.R
\docType{package}
\name{Rage-package}
\alias{Rage}
\alias{Rage-package}
\title{Rage: Life History Metrics from Matrix Population Models}
\description{
Functions for calculating life history metrics using matrix population models ('MPMs'). Described in Jones et al. (2021) \doi{10.1101/2021.04.26.441330}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonesor/Rage}
  \item Report bugs at \url{https://github.com/jonesor/Rage/issues}
}

}
\author{
\strong{Maintainer}: Owen Jones \email{jones@biology.sdu.dk} (\href{https://orcid.org/0000-0001-5720-4686}{ORCID})

Authors:
\itemize{
  \item Patrick Barks \email{barks@biology.sdu.dk} (\href{https://orcid.org/0000-0002-5947-8151}{ORCID})
  \item Pol Capdevila \email{pcapdevila.pc@gmail.com} (\href{https://orcid.org/0000-0002-2842-4302}{ORCID})
  \item Hal Caswell \email{h.caswell@uva.nl} (\href{https://orcid.org/0000-0003-4394-6894}{ORCID})
  \item Judy P. Che-Castaldo \email{jchecastaldo@lpzoo.org} (\href{https://orcid.org/0000-0002-9118-9202}{ORCID})
  \item Richard A. Hinrichsen \email{rich@hinrichsenenvironmental.com} (\href{https://orcid.org/0000-0003-0761-3005}{ORCID})
  \item John Jackson \email{jjackson0308@gmail.com} (\href{https://orcid.org/0000-0002-4563-2840}{ORCID})
  \item Tamora James \email{tamoradjames@protonmail.com} (\href{https://orcid.org/0000-0003-1363-4742}{ORCID})
  \item Sam Levin \email{levisc8@gmail.com} (\href{https://orcid.org/0000-0002-3289-9925}{ORCID})
  \item William K. Petry \email{wpetry@ncsu.edu} (\href{https://orcid.org/0000-0002-5230-5987}{ORCID})
  \item Roberto Salguero-Gomez \email{rob.salguero@zoo.ox.ac.uk} (\href{https://orcid.org/0000-0002-6085-4433}{ORCID})
  \item Iain Stott \email{stott@biology.sdu.dk} (\href{https://orcid.org/0000-0003-2724-7436}{ORCID})
  \item Chelsea C. Thomas \email{esochels@gmail.com} (\href{https://orcid.org/0000-0002-8155-9353}{ORCID})
  \item Christina M. Hernández \email{cmh352@cornell.edu} (\href{https://orcid.org/0000-0002-7188-8217}{ORCID})
  \item Lotte de Vries \email{c.devries@uva.nl} (\href{https://orcid.org/0000-0001-8955-0479}{ORCID})
  \item Stefano Giaimo \email{giaimo@evolbio.mpg.de} (\href{https://orcid.org/0000-0003-0421-3065}{ORCID})
}

Other contributors:
\itemize{
  \item Danny Buss \email{dlb50@cam.ac.uk} [contributor]
  \item Caroline Schuette \email{cschuette17@gmail.com} (\href{https://orcid.org/0000-0002-2063-8736}{ORCID}) [contributor]
}

}
\keyword{internal}
