% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizedvitalrates.R
\name{standardizedVitalRates}
\alias{standardizedVitalRates}
\title{Standardize vital rates}
\usage{
standardizedVitalRates(matU, matF, reproStages, matrixStages)
}
\arguments{
\item{matU}{Survival matrix}

\item{matF}{Fecundity matrix}

\item{reproStages}{Logical vector indicating which stages reproductive}

\item{matrixStages}{Character vector of matrix stage types (e.g. "propagule",
"active", or "dormant")}
}
\value{
A list (TODO:)
}
\description{
Standardize vital rates
}
\examples{
\dontrun{matU <- rbind( c(0, 0, 0, 0, 0), c(0.18, 0.16, 0, 0, 0), c(0.29, 0.23, 0.12,
0, 0), c(0, 0, 0.34, 0.53, 0), c(0, 0, 0, 0.87, 0) )

matF <- rbind( c(0, 0.13, 0, 0.96, 0), c(0, 0, 0, 0, 0), c(0, 0, 0, 0, 0),
c(0, 0, 0, 0, 0), c(0, 0, 0, 0, 0) )

reproStages <- c(FALSE, TRUE, FALSE, TRUE, FALSE)
matrixStages <- c('prop', 'active', 'active', 'active', 'active')

standardizedVitalRates(matU, matF, reproStages, matrixStages)}
}
\author{
Rob Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
