% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeMatrix.R
\name{standardizeMatrix}
\alias{standardizeMatrix}
\title{Transform a matrix population model to a standardized form}
\usage{
standardizeMatrix(matU, matF, matC = NULL, reproStages, matrixStages)
}
\arguments{
\item{matU}{A square matrix containing only survival-related transitions
(i.e. progression, stasis, retrogression).}

\item{matF}{A square matrix containing only sexual reproduction-related
transitions.}

\item{matC}{A square matrix containing only clonal reproduction-related
transitions. The default is \code{NULL}, indicating no clonal
reproduction (i.e. \code{matC} is a matrix of zeros).}

\item{reproStages}{Logical vector indicating which stages are reproductive}

\item{matrixStages}{Character vector of matrix stage types (e.g. "propagule",
"active", or "dormant")}
}
\value{
A list with four elements reflecting the standardized matrix, and
  it's components:
  \item{matA}{Standardized projection matrix}
  \item{matU}{Survival component of the standardized projection matrix}
  \item{matF}{Sexual reproduction component of the standardized projection matrix}
  \item{matC}{Clonal reproduction component of the standardized projection matrix}
}
\description{
Transform a matrix population model to a standardized set of stage classes
(e.g. propagule, pre-reproductive, reproductive, and post-reproductive). The
transition rates in the standardized matrix are a weighted average of the
transition rates from the relevant stages of the original matrix, weighted by
the relative proportion of each stage class expected at the stable
distribution.
}
\details{
This function is a wrapper for the functions
  \code{\link{rearrangeMatrix}}, \code{\link{reprodStages}} and
  \code{\link{collapseMatrix}}, which it calls in sequence.
}
\note{
The method used by this function to collapse a matrix population model
  preserves the equilibrium population growth rate (\eqn{lamda}) and relative
  stable distribution, but is not expected to preserve other traits such as
  relative reproductive values, sensitivities, net reproductive rates, life
  expectancy, etc.
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0,   0),
              c(  0, 0.3, 0.3, 0.1,   0),
              c(  0,   0, 0.4, 0.4, 0.1),
              c(  0,   0,   0, 0.1, 0.4))

matF <- rbind(c(  0, 1.1,   0, 1.6,   0),
              c(  0, 0.8,   0, 0.4,   0),
              c(  0,   0,   0,   0,   0),
              c(  0,   0,   0,   0,   0),
              c(  0,   0,   0,   0,   0))

matC <- rbind(c(  0, 0.6,   0, 0.5,   0),
              c(  0, 0.1,   0, 0.3,   0),
              c(  0,   0,   0,   0,   0),
              c(  0,   0,   0,   0,   0),
              c(  0,   0,   0,   0,   0))

reproStages <- c(FALSE, TRUE, FALSE, TRUE, FALSE)
matrixStages <- c('prop', 'active', 'active', 'active', 'active')

standardizeMatrix(matU, matF, matC, reproStages, matrixStages)

}
\author{
Rob Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
