% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_from_stage.R
\name{age_from_stage}
\alias{age_from_stage}
\alias{mpm_to_mx}
\alias{mpm_to_lx}
\alias{mpm_to_px}
\alias{mpm_to_hx}
\title{Calculate age-specific traits from a from matrix population model}
\usage{
mpm_to_mx(matU, matR, start = 1L, xmax = 1e+05, lxCrit = 1e-04)

mpm_to_lx(matU, start = 1L, xmax = 1e+05, lxCrit = 1e-04)

mpm_to_px(matU, start = 1L, xmax = 1e+05, lxCrit = 1e-04)

mpm_to_hx(matU, start = 1L, xmax = 1e+05, lxCrit = 1e-04)
}
\arguments{
\item{matU}{The survival component of a matrix population model (i.e. a
square projection matrix reflecting survival-related transitions; e.g.
progression, stasis, and retrogression)}

\item{matR}{The reproductive component of a matrix population model (i.e. a
square projection matrix reflecting transitions due to reproduction; either
sexual, clonal, or both)}

\item{start}{The index of the first stage at which the author considers the
beginning of life. Defaults to \code{1L}.}

\item{xmax}{Maximum age to which age-specific traits will be calculated
(defaults to \code{1e5}).}

\item{lxCrit}{Minimum value of lx to which age-specific traits will be
calculated (defaults to \code{1e-4}).}
}
\value{
A vector
}
\description{
These functions use age-from-stage decomposition methods to calculate
age-specific survivorship (lx), survival probability (px), mortality hazard
(hx), or reproduction (mx) from a matrix population model. A detailed
description of these methods can be found in sections 5.3.1 and 5.3.2 of
Caswell (2001).
}
\note{
The output vector is calculated recursively until the age class (x)
  reaches \code{xmax} or survivorship (lx) falls below \code{lxCrit} —
  whichever comes first. To force calculation to \code{xmax}, set
  \code{lxCrit = 0}. Conversely, to force calculation to \code{lxCrit}, set
  \code{xmax = Inf}.
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.4, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.4, 0.5))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

# age-specific survivorship
mpm_to_lx(matU)
mpm_to_lx(matU, start = 2)      # starting from stage 2
mpm_to_lx(matU, xmax = 10)      # to a maximum age of 10
mpm_to_lx(matU, lxCrit = 0.05)  # to a minimum lx of 0.05

# age-specific survival probability
mpm_to_px(matU)

# age-specific mortality hazard
mpm_to_hx(matU)

# age-specific fecundity
mpm_to_mx(matU, matF)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\seealso{
\code{\link{lifetable_convert}}
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>

Patrick Barks <patrick.barks@gmail.com>
}
