% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTableComponents.R
\name{ageSpecificRepro}
\alias{ageSpecificRepro}
\title{Calculate age-specific reproduction from stage-classified matrix model}
\usage{
ageSpecificRepro(matU, matR, startLife, N)
}
\arguments{
\item{matU}{A matrix containing only survival-related transitions (i.e.
progression, stasis, retrogression).}

\item{matR}{A matrix containing only reproduction-related transitions (either
sexual, clonal, or both; i.e. \code{matF}, \code{matC}, or \code{mat +
matC}).}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life.}

\item{N}{Maximum age to which age-specific reproduction will be calculated.}
}
\value{
A vector containing age-specific rates of reproduction, from ages 0
  through \code{N}.
}
\description{
This function uses age-from-stage decomposition methods to calculate
age-specific rates of reproduction from a stage-classified matrix population
model. A detailed description of these methods can be found in section 5.3.2
of Caswell (2001).
}
\examples{
matU <- rbind(c(0.0, 0.0, 0.0, 0.0),
              c(0.6, 0.6, 0.0, 0.0),
              c(0.2, 0.1, 0.3, 0.2),
              c(0.0, 0.2, 0.1, 0.1))

matF <- rbind(c(0.0, 0.0, 2.1, 3.6),
              c(0.0, 0.0, 0.3, 0.6),
              c(0.0, 0.0, 0.0, 0.0),
              c(0.0, 0.0, 0.0, 0.0))

ageSpecificRepro(matU, matF, startLife = 1, N = 20)
}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>
}
