% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeTableComponents.R
\name{ageSpecificSurv}
\alias{ageSpecificSurv}
\title{Calculate age-specific survivorship from matrix population model}
\usage{
ageSpecificSurv(matU, startLife, N)
}
\arguments{
\item{matU}{The survival component of a matrix population model (i.e. a
square projection matrix reflecting survival-related transitions; e.g.
progression, stasis, and retrogression)}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life.}

\item{N}{Maximum age to which age-specific survivorship will be calculated.}
}
\value{
A vector containing age-specific survivorship, from ages 0 through
  \code{N}.
}
\description{
This function uses age-from-stage decomposition methods to calculate
age-specific survivorship from a stage-classified matrix population model. A
detailed description of these methods can be found in section 5.3.1 of
Caswell (2001).
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

ageSpecificSurv(matU, startLife = 1, N = 20)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>
}
