% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEntropy.R
\name{dEntropy}
\alias{dEntropy}
\title{Calculate Demetrius' entropy}
\usage{
dEntropy(matU, matF, matC, startLife = 1, nSteps = 100)
}
\arguments{
\item{matU}{A square matrix containing only survival-related transitions
(i.e. progression, stasis, retrogression).}

\item{matF}{A square matrix containing only sexual reproduction-related
transitions.}

\item{matC}{A square matrix containing only clonal reproduction-related
transitions.}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life. Defaults to 1.}

\item{nSteps}{The age-cutoff for the decomposition of age-specific survival
(lx), sexual reproduction (mx), and/or clonal reproduction (cx).
This allows the user to exclude ages after which mortality or fertility has
plateaued (see function \code{qsdConverge} for more information). Defaults
to 100.}
}
\value{
A named list containing estimates of Demetrius' entropy with respect
  to sexual reproduction ('Fec', returned if \code{matF} is provided), clonal
  reproduction ('Clo', returned if \code{matC} is provided), and the sum of
  sexual and clonal reproduction ('FecClo', returned if both \code{matF} and
  \code{matC} are provided).
}
\description{
This function calculates Demetrius' entropy from a matrix population model,
by first using age-from-stage decomposition methods to estimate age-specific
survivorship (lx), fecundity (mx), and/or clonality (cx).
}
\examples{
matU <- rbind(c(0.0, 0.0, 0.0, 0.0),
              c(0.5, 0.0, 0.0, 0.0),
              c(0.0, 0.3, 0.0, 0.0),
              c(0.0, 0.0, 0.1, 0.1))

matF <- rbind(c(0.0, 0.0, 5.0, 10.0),
              c(0.0, 0.0, 0.0,  0.0),
              c(0.0, 0.0, 0.0,  0.0),
              c(0.0, 0.0, 0.0,  0.0))

matC <- rbind(c(0.0, 0.0, 0.0, 0.0),
              c(0.0, 0.0, 0.0, 1.0),
              c(0.0, 2.0, 0.0, 0.0),
              c(0.0, 0.0, 0.0, 0.0))

dEntropy(matU, matC = matC)
dEntropy(matU, matF, matC, nSteps = 5)
dEntropy(matU, matF, matC, nSteps = 10)
dEntropy(matU, matF, matC, nSteps = 100)
}
\references{
Demetrius, L. (1978) Adaptive value, entropy and survivorship
  curves. Nature 275, 213-214. doi:10.1038/275213a0
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
