% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dEntropy.R
\name{dEntropy}
\alias{dEntropy}
\title{Calculate Demetrius' entropy from a matrix population model}
\usage{
dEntropy(matU, matR, startLife = 1, nSteps = 100)
}
\arguments{
\item{matU}{A square matrix containing only survival-related transitions
(i.e. progression, stasis, retrogression).}

\item{matR}{A square matrix containing only reproduction-related transitions
(either sexual, clonal, or both; i.e. \code{matF}, \code{matC}, or
\code{matF + matC}).}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life. Defaults to 1.}

\item{nSteps}{The maximum age to which age-specific survival (lx) and
reproduction (mx) will be calculated. This allows the user to exclude ages
after which mortality or fertility has plateaued (see function
\code{\link{qsdConverge}} for more information). Defaults to 100.}
}
\value{
Demetrius' entropy.
}
\description{
Calculates Demetrius' entropy from a matrix population model, by first using
age-from-stage decomposition methods to estimate age-specific survivorship
(lx) and fecundity (mx).
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

dEntropy(matU, matF, nSteps = 10)
dEntropy(matU, matF, nSteps = 100)

}
\references{
Demetrius, L. (1978) Adaptive value, entropy and survivorship
  curves. Nature 275, 213-214. doi:10.1038/275213a0
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Patrick Barks <patrick.barks@gmail.com>
}
