% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_d.R
\name{entropy_d}
\alias{entropy_d}
\title{Calculate Demetrius' entropy from trajectories of age-specific survivorship
and fecundity}
\usage{
entropy_d(lx, mx)
}
\arguments{
\item{lx}{Age-specific survivorship trajectory (a vector of
monotonically-declining values in the interval [0,1])}

\item{mx}{Age-specific fecundity trajectory (a vector of non-negative values)}
}
\value{
Demetrius' entropy.
}
\description{
This function calculates Demetrius' entropy from vectors of age-specific
survivorship (lx) and fecundity (mx).
}
\examples{
lx <- c(1.00, 0.85, 0.70, 0.65, 0.55, 0.50, 0.45, 0.40, 0.35)
mx <- c(0.00, 0.00, 1.10, 1.50, 1.60, 1.70, 1.50, 1.20, 0.70)

entropy_d(lx, mx)

}
\references{
Demetrius, L. (1978) Adaptive value, entropy and survivorship
  curves. Nature 275, 213-214. doi:10.1038/275213a0
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Patrick Barks <patrick.barks@gmail.com>
}
