% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_k2.R
\name{entropy_k2}
\alias{entropy_k2}
\title{Calculate Keyfitz entropy for a matrix population model}
\usage{
entropy_k2(A, type = "age")
}
\arguments{
\item{A}{A square numeric projection matrix The dimension of the matrix
corresponds to the number of age classes}

\item{type}{An argument specifying the type of matrix model used in the
calculations. This is necessary because the calculations for age vs.
stage-based matrices are different. Possible options are `age` and `stage`.}
}
\value{
Returns a single numeric value representing the Keyfitz entropy
for the given matrix. This value quantifies the dispersion of age at death.
}
\description{
This function computes Keyfitz entropy from a transition matrix (the A
matrix). The function is designed to work with discrete, age-classified
matrix models.
}
\examples{
# Define a matrix population model for a hypothetical population
A <- matrix(c(
  0.7, 0.0, 0.0, 0.5,
  0.4, 0.8, 0.0, 0.0,
  0.0, 0.2, 0.9, 0.0,
  0.0, 0.0, 0.2, 0.1
), nrow = 4, byrow = TRUE)

# Calculate Keyfitz entropy
entropy_k2(A)

}
\references{
Keyfitz, N. 1977. Applied Mathematical Demography. New York:
  Wiley.
  
  de Vries, C., Bernard, C., & Salguero-Gómez, R. 2023. Discretising
  Keyfitz' entropy for studies of actuarial senescence and comparative
  demography. Methods in Ecology and Evolution, 14, 1312–1319.
  https://doi.org/10.1111/2041-210X.14083
}
\seealso{
Other life history traits: 
\code{\link{entropy_d}()},
\code{\link{entropy_k}()},
\code{\link{gen_time}()},
\code{\link{life_expect_mean}()},
\code{\link{longevity}()},
\code{\link{net_repro_rate}()},
\code{\link{repro_maturity}},
\code{\link{shape_rep}()},
\code{\link{shape_surv}()}
}
\author{
Lotte de Vries <c.devries@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>
}
\concept{life history traits}
