% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longevity.R
\name{longevity}
\alias{longevity}
\title{Calculate measures of life expectancy and longevity}
\usage{
longevity(matU, startLife = 1, initPop = 100, run = 1000)
}
\arguments{
\item{matU}{A matrix containing only survival-dependent processes (growth,
stasis, shrinkage).}

\item{startLife}{The first stage at which the author consider the beginning
of life.}

\item{initPop}{Initial population size.}

\item{run}{Number of iterations that maximum longevity will be calculated
over.}
}
\value{
Returns a list including the following named elements (each of which
  contain a single numeric value):

  - \code{eta}: mean life expectancy conditional on entering the life cycle
  at life stage described by \code{startLife}

  - \code{var_eta}: variance in life expectancy conditional on entering the
  life cycle at life stage described by \code{startLife}

  - \code{Max}: maximum longevity observed by iterating a population vector
  with \code{initPop} individuals in stage \code{startLife} up to \code{run}
  times
}
\description{
A function to calculate life expectancy and maximum longevity of individuals
in a matrix population model
}
\details{
This function applies Markov chain approaches to obtain the mean and variance
in life expectancy, and a loop to calculate maximum longevity.
}
\note{
%% ~~further notes~~
}
\examples{
matU <- matrix(c(0, 0, 0, 0, 0.5, 0, 0, 0, 0, 0.3, 0, 0, 0, 0, 0.1, 0.1),
               nrow = 4, byrow = TRUE)

longevity(matU, startLife = 1, initPop = 100, run = 1000)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968

  Morris, W. F., and D. F. Doak. 2003. Quantitative Conservation Biology:
  Theory and Practice of Population Viability Analysis. Sinauer Associates,
  Sunderland, Massachusetts, USA
}
\author{
Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <hcaswell@whoi.edu>
}
\keyword{~kwd1}
\keyword{~kwd2}
