% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLifeTable.R
\name{makeLifeTable}
\alias{makeLifeTable}
\title{Produces a life table from a matrix population model}
\usage{
makeLifeTable(matU, matF = NULL, matC = NULL, startLife = 1,
  nSteps = 1000)
}
\arguments{
\item{matU}{A matrix containing only survival-dependent processes (e.g.
progression, stasis, retrogression).}

\item{matF}{A matrix containing only sexual reproduction, with zeros
elsewhere.}

\item{matC}{A matrix containing only clonal reproduction, with zeros
elsewhere. If not provided, it defaults to a matrix with all zeros.}

\item{startLife}{The first stage at which the author considers the beginning
of life in the life cycle of the species. It defaults to the first stage.}

\item{nSteps}{Number of time steps for which the life table will be
calculated. This is on the same units as the matrix population model - see
MatrixPeriodicity in metadata of COMPADRE/COMADRE.}
}
\value{
A `data.frame` with between 5 and 8 columns. If just `matU` is
  provided, the `data.frame` has columns `x` (age), `lx` (survivorship), 'dx'
  (proportion of the original cohort dying during each stage), 'qx' (force of
  mortality), and 'ex' (age-specific life expectancy). If `matF` is provided
  (in addition to `matU`), `mx` (age-specific sexual reproduction) and lxmx
  (# sexually-produced recruits per capita in each age) are included in the
  output. If `matC` is provided (in addition to `matU`), `cx` (age-specific
  clonal reproduction) and lxcx (# clonally-produced recruits per capita in
  each age) are included in the output.
}
\description{
This function uses age-from-stage decompositions to calculate a life table
from a matrix population model
}
\details{
A detailed description of these methods can be found in section 5.3 of
Caswell (2001) and the supplementary information of Jones et al. (2014).
}
\note{
%% ~~further notes~~
}
\examples{


matU <- matrix(c(0, 0, 0, 0, 0.5, 0, 0, 0, 0, 0.3, 0, 0, 0, 0, 0.1, 0.1), nrow = 4, byrow = TRUE)
matF <- matrix(c(0, 0, 5, 10, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0.), nrow = 4, byrow = TRUE)
matC <- matrix(c(0, 0, 0, 0, 0, 1, 2, 3, 0, 0, 0, 0, 0, 0, 0, 0), nrow = 4, byrow = TRUE)

makeLifeTable(matU, matF, matC, startLife = 1, nSteps = 100)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction, Analysis, and
Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Caswell, H. (2006) Applications of Markov chains in demography. pp. 319-334
in A.N. Langville and W.J. Stewart (editors) MAM2006: Markov Anniversary
Meeting. Boson Books, Raleigh, North Caroline, USA

Jones, O.R. et al. (2014) Diversity of ageing across the tree of life.
Nature, 505(7482), 169–173
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>
}
\keyword{~kwd1}
\keyword{~kwd2}
