% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLifeTable.R
\name{makeLifeTable}
\alias{makeLifeTable}
\title{Generate a life table from a matrix population model}
\usage{
makeLifeTable(matU, matF = NULL, matC = NULL, startLife = 1,
  nSteps = 1000)
}
\arguments{
\item{matU}{The survival component of a matrix population model (i.e. a
square projection matrix reflecting survival-related transitions; e.g.
progression, stasis, and retrogression)}

\item{matF}{(Optional) The sexual component of a matrix population model
(i.e. a square projection matrix reflecting transitions due to sexual
reproduction)}

\item{matC}{(Optional) The clonal component of a matrix population model
(i.e. a square projection matrix reflecting transitions due to clonal
reproduction).}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life. Defaults to 1.}

\item{nSteps}{Number of time steps for which the life table will be
calculated. Time steps are in the same units as the matrix population model
(see MatrixPeriodicity column in metadata of COM(P)ADRE). Defaults to 1000.}
}
\value{
A \code{data.frame} containing 7-13 columns and \code{nSteps} rows.
  Columns include:
  \item{x}{age}
  \item{lx}{survivorship to start of age class x}
  \item{dx}{proportion of original cohort dying in interval [x, x+1)}
  \item{qx}{force of mortality at age x}
  \item{Lx}{survivorship to middle of age class x}
  \item{Tx}{proportion of original cohort alive at age x and beyond}
  \item{ex}{remaining life expectancy at age x}

If \code{matF} is provided, also includes:
  \item{mx}{per-capita rate of sexual reproduction at age x}
  \item{lxmx}{expected number of sexual offspring per original
  cohort member produced at age x}
  
If \code{matC} is provided, also includes:
  \item{cx}{per-capita rate of clonal reproduction at age x}
  \item{lxcx}{expected number of clonal offspring per original
  cohort member produced at age x}
  
If both \code{matF} and \code{matC} are provided, also includes:
  \item{mxcx}{per-capita rate of total reproduction (sexual + clonal) at age x}
  \item{lxmxcx}{expected number of total offspring (sexual + clonal) per original
  cohort member produced at age x}
}
\description{
This function uses age-from-stage decomposition methods to generate a life
table from a matrix population model. A detailed description of these methods
can be found in section 5.3 of Caswell (2001).
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))
              
matC <- rbind(c(  0,   0, 0.4, 0.5),
              c(  0,   0, 0.3, 0.1),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

makeLifeTable(matU, startLife = 1, nSteps = 100)
makeLifeTable(matU, matF, startLife = 1, nSteps = 100)
makeLifeTable(matU, matF, matC, startLife = 1, nSteps = 100)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction, Analysis, and
Interpretation. Sinauer Associates; 2nd edition. ISBN: 978-0878930968

Caswell, H. (2006) Applications of Markov chains in demography. pp. 319-334
in A.N. Langville and W.J. Stewart (editors) MAM2006: Markov Anniversary
Meeting. Boson Books, Raleigh, North Caroline, USA

Jones, O.R. et al. (2014) Diversity of ageing across the tree of life.
Nature, 505(7482), 169–173
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <h.caswell@uva.nl>

Owen R. Jones <jones@biology.sdu.dk>
}
