% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rage-package.R
\docType{package}
\name{Rage-package}
\alias{Rage-package}
\title{\pkg{Rage}}
\description{
An R package with tools to derive life history traits from matrix population
models.
}
\references{
Salguero-Gomez, R., Jones, O. R., Archer, C. R., Buckley, Y. M.,
Che-Castaldo, J., Caswell, H., Hodgson, D., Scheuerlein, A., Conde, D. A.,
Brinks, E., de Buhr, H., Farack, C., Gottschalk, F., Hartmann, A., Henning,
A., Hoppe, G., Roemer, G., Runge, J., Ruoff, T., Wille, J., Zeh, S., Davison,
R., Vieregg, D., Baudisch, A., Altwegg, R., Colchero, F., Dong, M., de Kroon,
H., Lebreton, J.-D., Metcalf, C. J. E., Neel, M. M., Parker, I. M., Takada,
T., Valverde, T., Velez-Espino, L. A., Wardle, G. M., Franco, M., Vaupel, J.
W. (2015) The COMPADRE Plant Matrix Database: an open online repository for
plant demography. Journal of Ecology, 103: 202-218. 
\url{https://doi.org/10.1111/1365-2745.12334}

Salguero-Gomez, R., Jones, O. R., Archer, C. R., Bein, C., de Buhr, H.,
Farack, C., Gottschalk, F., Hartmann, A., Henning, A., Hoppe, G., Roemer, G.,
Ruoff, T., Sommer, V., Wille, J., Voigt, J., Zeh, S., Vieregg, D., Buckley,
Y. M., Che-Castaldo, J., Hodgson, D., Scheuerlein, A., Caswell, H., Vaupel,
J. W. (2016) COMADRE: a global data base of animal demography. Journal of
Animal Ecology, 85: 371-384. \url{https://doi.org/10.1111/1365-2656.12482}
}
\author{
Roberto Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>

Owen R. Jones <jones@biology.sdu.dk>
}
