% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro_maturity.R
\name{repro_maturity}
\alias{repro_maturity}
\alias{mature_prob}
\alias{mature_age}
\alias{mature_life_expect}
\title{Calculate traits relating to the age of reproductive maturity from a matrix
population model}
\usage{
mature_prob(matU, matR, startLife = 1)

mature_age(matU, matR, startLife = 1)

mature_life_expect(matU, matR, startLife = 1)
}
\arguments{
\item{matU}{The survival component of a matrix population model (i.e. a
square projection matrix reflecting survival-related transitions; e.g.
progression, stasis, and retrogression)}

\item{matR}{The reproductive component of a matrix population model (i.e. a
square projection matrix reflecting transitions due to reproduction; either
sexual, clonal, or both)}

\item{startLife}{The index of the first stage at which the author considers
the beginning of life. Defaults to 1.}
}
\value{
Trait value (scalar)
}
\description{
Apply Markov chain approaches to decompose moments along the age-specific
trajectory of reproduction for individuals in a matrix population model.
Includes functions to calculated the probability of achieving reproductive
maturity (\code{mature_prob}), mean age at first reproduction
(\code{mature_age}), and remaining life expectancy at the age of first
reproduction (\code{mature_life_expect}).
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

mature_prob(matU, matF, startLife = 1)
mature_age(matU, matF, startLife = 1)
mature_life_expect(matU, matF, startLife = 1)

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
  Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
  978-0878930968
}
\author{
Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>

Hal Caswell <hcaswell@whoi.edu>

Owen R. Jones <jones@biology.sdu.dk>

Patrick Barks <patrick.barks@gmail.com>
}
