% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodStages.R
\name{reprodStages}
\alias{reprodStages}
\title{Group the stages of a matrix population model into a standardized set of
stage classes}
\usage{
reprodStages(matF, reproStages, matrixStages, includeProp = TRUE,
  includePost = TRUE)
}
\arguments{
\item{matF}{(matrix) fecundity matrix, rearranged so that non-reproductive
stages are in the final rows/columns}

\item{reproStages}{Logical vector identifying which stages reproductive}

\item{matrixStages}{(character) vector of stages, values are "prop"
(propagule), "active", and "dorm" (dormant)}

\item{includeProp}{(logical) include propagule stage. default: \code{TRUE}.
if \code{TRUE}, propagule stage (if present) is given back in result. If
\code{FALSE}, it's included into the pre-reproductive stage}

\item{includePost}{(logical) include post-reproductive stage. default:
\code{TRUE}. if \code{TRUE}, post-reproductive stage (if present) is given
back in result. If \code{FALSE}, it's included into the reproductive
stage}
}
\description{
Assumes that fecundity and mean fecundity matrices have been rearranged so
that non-reproductive stages are in the final rows/columns.
}
\details{
Output indicates groupings to be used when collapsing the matrix model.
}
\note{
Dormant stages are not currently handled.
}
\examples{
matF <- rbind(c(  0, 1.1,   0, 1.6,   0),
              c(  0, 0.8,   0, 0.4,   0),
              c(  0,   0,   0,   0,   0),
              c(  0,   0,   0,   0,   0),
              c(  0,   0,   0,   0,   0))

reproStages <- c(FALSE, TRUE, FALSE, TRUE, FALSE)
matrixStages <- c('prop', 'active', 'active', 'active', 'active')
reprodStages(matF, reproStages, matrixStages)

# combine post-reproductive and reproductive
reprodStages(matF, reproStages, matrixStages, includePost = FALSE)
}
\seealso{
\code{\link{standardizeMatrix}}
}
\author{
Rob Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
