% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMatrix.R
\name{splitMatrix}
\alias{splitMatrix}
\title{Splits a matrix population model into the constituent U, F and C matrices}
\usage{
splitMatrix(matA)
}
\arguments{
\item{matA}{A matrix population model (i.e. a square projection matrix)}
}
\value{
A list of three matrices: \code{matU},\code{matF} and \code{matC}
}
\description{
This function splits a matrix population model into three constituent
matrices, U (growth and survival processes), F (sexual reproduction) and C
(clonal reproduction). Warning! The functionality is very basic — it assumes
that sexual reproduction is located in the top row of the matrix, and that
everything else is growth or survival (the U matrix). Clonality is assumed to
be non-existent.
}
\examples{
matA <- rbind(c(0.1,   0, 5.3, 4.2),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

splitMatrix(matA)

}
\author{
Owen R. Jones <jones@biology.sdu.dk>
}
