% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitalRatePerturbation.R
\name{vitalRatePerturbation}
\alias{vitalRatePerturbation}
\title{Perturbation analysis of vital rates in a matrix population model}
\usage{
vitalRatePerturbation(matU, matF, matC = NULL, pert = 1e-06,
  demogstat = "lambda", ...)
}
\arguments{
\item{matU}{A square matrix containing only survival-related transitions
(i.e. progression, stasis, retrogression).}

\item{matF}{A square matrix containing only sexual reproduction-related
transitions.}

\item{matC}{A square matrix containing only clonal reproduction-related
transitions.}

\item{pert}{Magnitude of the perturbation (defaults to \code{1e-6}).}

\item{demogstat}{The demographic statistic to be used, as in "the
sensitivity/elasticity of ___ to vital rate perturbations." Defaults to the
per-capita population growth rate at equilibrium (\eqn{\lambda}). Also
accepts a user-supplied function that performs a calculation on a
projection matrix and returns a single numeric value.}

\item{...}{Additional arguments passed to the function \code{demogstat}.}
}
\value{
A data frame with 1 row and 10 columns:
\item{SSurvival}{sensitivity of \code{demogstat} to survival}
\item{SGrowth}{sensitivity of \code{demogstat} to growth}
\item{SShrinkage}{sensitivity of \code{demogstat} to shrinkage}
\item{SReproduction}{sensitivity of \code{demogstat} to sexual reproduction}
\item{SClonality}{sensitivity of \code{demogstat} to clonality}
\item{ESurvival}{elasticity of \code{demogstat} to survival}
\item{EGrowth}{elasticity of \code{demogstat} to growth}
\item{EShrinkage}{elasticity of \code{demogstat} to shrinkage}
\item{EReproduction}{elasticity of \code{demogstat} to sexual reproduction}
\item{EClonality}{elasticity of \code{demogstat} to clonality}
}
\description{
Perturbs lower-level vital rates within a matrix population model and
measures the response of the per-capita population growth rate at equilibrium
(\eqn{lambda}), or, with a user-supplied function, any other demographic
statistic.
}
\examples{

\dontrun{
data(Compadre)
pira <- subsetDB(Compadre, SpeciesAccepted == "Pinus radiata")

vitalRatePerturbation(pira)  # Not implemented yet!

vitalRatePerturbation(matU = pira@matU, matF = pira@matF)
vitalRatePerturbation(matU = pira@matU, matF = pira@matF, pert = 1e-03)
# use a smaller perturbation
vitalRatePerturbation(matU = pira@matU, matF = pira@matF, pert = 1e-10)

# calculate the sensitivity/elasticity of the damping ratio to vital rate perturbations
damping <- function(matU, matF, matC){  # define function for damping ratio
A <- matU+matF+matC
eig <- eigen(A)$values
dm <- rle(Mod(eig))$values
return(dm[1] / dm[2])
}
vitalRatePerturbation(matU = pira@matU, matF = pira@matF, stat = damping)
vitalRatePerturbation(matU = pira@matU, matF = pira@matF, stat = "damping")
}

}
\author{
Rob Salguero-Gómez <rob.salguero@zoo.ox.ac.uk>
}
