% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitalRates.R
\name{vitalRates}
\alias{vitalRates}
\title{Derive mean vital rates from a matrix population model}
\usage{
vitalRates(matU, matF, matC = NULL, splitStages = FALSE,
  weighted = FALSE)
}
\arguments{
\item{matU}{A square matrix containing only survival-related transitions
(i.e. progression, stasis, retrogression).}

\item{matF}{A square matrix containing only sexual reproduction-related
transitions.}

\item{matC}{A square matrix containing only clonal reproduction-related
transitions.}

\item{splitStages}{Splits vital rates according to some pre-determined
criteria (below).}

\item{weighted}{Allows to weight mean vital rates according to various
criteria (below).}
}
\value{
- 'Weighted': This argument allows to weight mean values of vital
rates (survival 'surv', progression 'prog', retrogression 'retr', sexual
reproduction 'fec' and clonal reproduction 'clo') with an equal contribution
for all stages (default), by the stable st/age distribution ('SSD'), or by a
given population vector chosen by the user, so long as it is congruent with
the dimensions of the chosen 'matU', 'matF', and 'matC'.

- 'splitStages': This argument allows to split the values of vital rates
according to recognizable stages in the matrix. When 'all', all vital rates
are averaged all existing stages, if 'ontogeny', they are averaged as
juveniles ('Juv') and adults ('Adu'), and if by 'MatrixClassOrganized', it
takes a vector with the pre-established stages of
'compadre$matrixClass[[i]]$MatrixClassOrganized' or
'compadre$matrixClass[[i]]$MatrixClassOrganized', where 'i' is the index of
the chosen study in 'COMPADRE' or 'COMADRE'.
}
\description{
Derive mean vital rates from a matrix population model corresponding to
separate demographic processes. Specifically, this function decomposes vital
rates of survival, progression, retrogression, sexual reproduction and clonal
reproduction according to various ways of weighted means and organization of
stages along the life cycle represented in the matrix population model.
}
\examples{
matU <- rbind(c(0.1,   0,   0,   0),
              c(0.5, 0.2, 0.1,   0),
              c(  0, 0.3, 0.3, 0.1),
              c(  0,   0, 0.5, 0.6))

matF <- rbind(c(  0,   0, 1.1, 1.6),
              c(  0,   0, 0.8, 0.4),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))
              
matC <- rbind(c(  0,   0, 0.4, 0.5),
              c(  0,   0, 0.3, 0.1),
              c(  0,   0,   0,   0),
              c(  0,   0,   0,   0))

#Vital rate outputs without weights:
vitalRates(matU, matF, matC, splitStages = 'all', weighted = FALSE)
vitalRates(matU, matF, matC, splitStages = 'ontogeny', weighted = FALSE)
vitalRates(matU, matF, matC,
           splitStages = c('prop', 'active', 'active', 'active'),
           weighted = FALSE)


#Vital rate outputs weighted by the stable stage distribution of 'matA':
vitalRates(matU, matF, matC, splitStages = 'all', weighted = 'SSD')
vitalRates(matU, matF, matC, splitStages = 'ontogeny', weighted = 'SSD')
vitalRates(matU, matF, matC,
           splitStages = c('prop', 'active', 'active', 'active'),
           weighted = 'SSD')

}
\references{
Caswell, H. (2001) Matrix Population Models: Construction,
Analysis, and Interpretation. Sinauer Associates; 2nd edition. ISBN:
978-0878930968
}
\author{
Roberto Salguero-Gomez <rob.salguero@zoo.ox.ac.uk>
}
